Add tests for @builtin.get_commands
Which currently fail because it has not been updated to the
thingpedia-client interface.
